requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
backoff==2.2.1
bcrypt==4.2.0
bleach==6.2.0
blinker==1.8.2
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.17
click==8.1.7
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
Deprecated==1.2.14
docutils==0.21.2
durationpy==0.9
executing==2.1.0
fastapi==0.115.4
filelock==3.16.1
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.10.0
google-auth==2.35.0
googleapis-common-protos==1.65.0
grpcio==1.67.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.29.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
kaggle==1.6.17
kagglehub==0.3.3
keyring==25.5.0
kiwisolver==1.4.7
kubernetes==31.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplot==0.1.9
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
more-itertools==10.5.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
nh3==0.2.18
numpy==2.1.2
oauthlib==3.2.2
onnxruntime==1.20.0
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.11
overrides==7.7.0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
pkginfo==1.10.0
platformdirs==4.3.6
posthog==3.7.0
prompt_toolkit==3.0.48
protobuf==4.25.5
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyloco==0.0.139
pyparsing==3.2.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
readme_renderer==44.0
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
rsa==4.9
safetensors==0.4.5
setuptools==75.3.0
shellingham==1.5.4
SimpleWebSocketServer==0.1.2
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.41.2
sympy==1.13.1
tenacity==9.0.0
text-unidecode==1.3
tokenizers==0.20.1
torch==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.66.6
traitlets==5.14.3
transformers==4.46.1
twine==5.1.1
typer==0.12.5
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
ushlex==0.99.1
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.1.1
wrapt==1.16.0
zipp==3.20.2
